+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Currently in development.
 - Added optional support for background tasks
 - Allow ftp, file, sftp and smb scheme and IPv6 addresses in URLs
 - Improve display of quantities
+- Improve static file caching
 
 Version 0.20
 ------------